update accessor functions to be consistent with branch types

+++ include/Muon.h
@@ -31,7 +31,7 @@ class Muon : public Object {
   float CocktailPt              ();
   float CocktailPtError         ();
 
-  float   Charge                  ();
+  int   Charge                  ();
 
 
   // IDs
